+++ R/plotting-Spectrum1.R
@@ -1,21 +1,21 @@
 plot_Spectrum1 <- function(spectrum,
-                           centroided.=FALSE,
-                           plot=TRUE,
-                           w1) {
+                           centroided. = centroided(spectrum),
+                           plot = TRUE) {
+  if (is.na(centroided.)) {
+    centroided. <- TRUE
+    message("No centroided/profile information. Setting to centroided")
+  }
   mtc <- mz(spectrum)
   i <- intensity(spectrum)
+  dfr <- data.frame(i = i, mtc = mtc)
   if (centroided.) {
-    if (missing(w1))
-      w1 <- max(mtc)/500
-    dfr <- data.frame(i = i, mtc = mtc, width = w1)
-    p <- ggplot(dfr, aes(x = mtc, y = i, width = width)) +
-      geom_bar(stat="identity", position="identity")
+   p <- ggplot(dfr, aes(x = mz, xend = mz, y = 0, yend = i)) +
+      geom_segment()
   } else {
-    dfr <- data.frame(i = i, mtc = mtc)
     p <- ggplot(dfr, aes(x = mtc, y = i)) + geom_line()
   }
   title <- ggtitle(paste("Retention time", rtime(spectrum)))
-  p <- p + labs(x="M/Z", y="Intensity") + title
+  p <- p + labs(x = "M/Z", y = "Intensity") + title
   if (plot)
     print(p + title)
   invisible(p + title)